+++ .github/upload-preview.js
@@ -1,13 +1,17 @@
+const fs = require('fs');
 const path = require('path');
 const { Octokit } = require('@octokit/rest');
 const octokit = new Octokit({ auth: process.env.GH_PR_TOKEN });
 const surge = require('surge');
 const publishFn = surge().publish();
 
-const owner = process.env.CIRCLE_PROJECT_USERNAME; // patternfly
-const repo = process.env.CIRCLE_PROJECT_REPONAME;
-const prnum = process.env.CIRCLE_PR_NUMBER;
-const prbranch = process.env.CIRCLE_BRANCH;
+// From github actions
+const ghrepo = process.env.GITHUB_REPOSITORY || '';
+
+const owner = process.env.CIRCLE_PROJECT_USERNAME || ghrepo.split('/')[0]; // patternfly
+const repo = process.env.CIRCLE_PROJECT_REPONAME || ghrepo.split('/')[1];
+const prnum = process.env.CIRCLE_PR_NUMBER || process.env.GH_PR_NUM;
+const prbranch = process.env.CIRCLE_BRANCH || process.env.GITHUB_REF.split('/').pop();
 
 const uploadFolder = process.argv[2];
 if (!uploadFolder) {
@@ -16,24 +20,20 @@ if (!uploadFolder) {
 }
 
 const uploadFolderName = path.basename(uploadFolder);
-let uploadURL = `${repo}-${prnum ? `pr-${prnum}` : prbranch}`
-  .replace(/[\/|\.]/g, '-')
-  .replace('-main', '');
+let uploadURL = `${repo}-${prnum ? `pr-${prnum}` : prbranch}`.replace(/[\/|\.]/g, '-');
 
-if (uploadFolderName === '.out') {
-  uploadURL += '-pf3';
+switch(uploadFolderName) {
+  case 'coverage':
+    uploadURL += '-a11y.surge.sh';
+    break;
+  case 'public':
+    uploadURL += '.surge.sh';
+    break;
+  default:
+    uploadURL += `-${uploadFolderName}`;
+    uploadURL += '.surge.sh';
+    break;
 }
-else if (uploadFolderName === 'dist') {
-  uploadURL += `-a11y`;
-}
-else if (uploadFolderName === 'results') {
-  uploadURL += `-cypress`;
-}
-else if (uploadFolderName !== 'public') {
-  uploadURL += `-${uploadFolderName}`;
-}
-
-uploadURL += '.surge.sh';
 
 publishFn({
   project: uploadFolder,
@@ -63,20 +63,14 @@ if (prnum) {
       const existingComment = comments.find(comment => comment.user.login === 'patternfly-build');
       if (existingComment) {
         commentBody += existingComment.body.trim();
-        commentBody += '\n';
+        commentBody += '\n\n';
       }
 
-      if (uploadFolderName === '.out') {
-        commentBody += tryAddComment(`PF3 preview: https://${uploadURL}`, commentBody);
-      }
-      else if (uploadFolderName === 'public') {
-        commentBody += tryAddComment(`PF4 preview: https://${uploadURL}`, commentBody);
+      if (uploadFolderName === 'public') {
+        commentBody += tryAddComment(`Preview: https://${uploadURL}`, commentBody);
       }
       else if (uploadFolderName === 'dist') {
         commentBody += tryAddComment(`A11y report: https://${uploadURL}`, commentBody);
-      }
-      else if (uploadFolderName === 'report') {
-        commentBody += tryAddComment(`Cypress report: https://${uploadURL}`, commentBody);
       }
 
       if (existingComment) {