+++ packages/react-core/src/components/Menu/MenuContext.ts
@@ -15,6 +15,7 @@ export const MenuContext = React.createContext<{
   flyoutRef?: React.Ref<HTMLLIElement>;
   setFlyoutRef?: (ref: React.Ref<HTMLLIElement>) => void;
   disableHover?: boolean;
+  isScrollable?: boolean;
 }>({
   menuId: null,
   parentMenu: null,
@@ -29,7 +30,8 @@ export const MenuContext = React.createContext<{
   onGetMenuHeight: () => null,
   flyoutRef: null,
   setFlyoutRef: () => null,
-  disableHover: false
+  disableHover: false,
+  isScrollable: false
 });
 
 export const MenuItemContext = React.createContext<{