+++ packages/react-core/src/components/Modal/__tests__/ModalContent.test.tsx
@@ -3,9 +3,14 @@ import { shallow } from 'enzyme';
 
 import { ModalContent } from '../ModalContent';
 
+const modalContentProps = {
+  boxId: "boxId",
+  labelId: "labelId",
+  descriptorId: "descriptorId"
+}
 test('Modal Content Test only body', () => {
   const view = shallow(
-    <ModalContent title="Test Modal Content title" id="id" isOpen>
+    <ModalContent title="Test Modal Content title" isOpen {...modalContentProps}>
       This is a ModalBox header
     </ModalContent>
   );
@@ -14,7 +19,7 @@ test('Modal Content Test only body', () => {
 
 test('Modal Content Test isOpen', () => {
   const view = shallow(
-    <ModalContent title="Test Modal Content title" id="id" isOpen>
+    <ModalContent title="Test Modal Content title" isOpen {...modalContentProps}>
       This is a ModalBox header
     </ModalContent>
   );
@@ -23,7 +28,7 @@ test('Modal Content Test isOpen', () => {
 
 test('Modal Content Test description', () => {
   const view = shallow(
-    <ModalContent title="Test Modal Content title" id="id" isOpen description="This is a test description.">
+    <ModalContent title="Test Modal Content title" isOpen description="This is a test description." {...modalContentProps}>
       This is a ModalBox header
     </ModalContent>
   );
@@ -32,7 +37,7 @@ test('Modal Content Test description', () => {
 
 test('Modal Content Test with footer', () => {
   const view = shallow(
-    <ModalContent title="Test Modal Content title" id="id" isOpen actions={['Testing']}>
+    <ModalContent title="Test Modal Content title" isOpen actions={['Testing']} {...modalContentProps}>
       This is a ModalBox header
     </ModalContent>
   );
@@ -41,7 +46,7 @@ test('Modal Content Test with footer', () => {
 
 test('Modal Content test without footer', () => {
   const view = shallow(
-    <ModalContent title="Test Modal Content title" id="id" isOpen>
+    <ModalContent title="Test Modal Content title" isOpen {...modalContentProps}>
       This is a ModalBox header
     </ModalContent>
   );
@@ -53,10 +58,10 @@ test('Modal Content Test with onclose', () => {
     <ModalContent
       title="Test Modal Content title"
       actions={['Testing footer']}
-      isLarge
+      variant="large"
       onClose={() => undefined}
-      id="id"
       isOpen
+      {...modalContentProps}
     >
       This is a ModalBox header
     </ModalContent>
@@ -72,10 +77,10 @@ test('Modal Test with custom header', () => {
       header={header}
       title="test-custom-header-modal"
       actions={['Testing footer']}
-      isLarge
+      variant="large"
       onClose={() => undefined}
-      id="id"
       isOpen
+      {...modalContentProps}
     >
       This is a ModalBox header
     </ModalContent>
@@ -87,7 +92,7 @@ test('Modal Test with custom footer', () => {
   const footer = <span id="test-custom-footer">TEST</span>;
 
   const view = shallow(
-    <ModalContent footer={footer} title="Test Modal Custom Footer" isLarge onClose={() => undefined} id="id" isOpen>
+    <ModalContent footer={footer} title="Test Modal Custom Footer" variant="large" onClose={() => undefined} isOpen {...modalContentProps}>
       This is a ModalBox header
     </ModalContent>
   );