+++ packages/react-core/src/components/Progress/Progress.tsx
@@ -11,7 +11,7 @@ export enum ProgressSize {
   lg = 'lg'
 }
 
-export interface ProgressProps extends Omit<React.HTMLProps<HTMLDivElement>, 'size' | 'label'> {
+export interface ProgressProps extends Omit<React.HTMLProps<HTMLDivElement>, 'size' | 'label' | 'title'> {
   /** Classname for progress component. */
   className?: string;
   /** Size variant of progress. */
@@ -20,8 +20,8 @@ export interface ProgressProps extends Omit<React.HTMLProps<HTMLDivElement>, 'si
   measureLocation?: 'outside' | 'inside' | 'top' | 'none';
   /** Status variant of progress. */
   variant?: 'danger' | 'success' | 'warning';
-  /** Title above progress. */
-  title?: string;
+  /** Title above progress. The isTitleTruncated property will only affect string titles. Node title truncation must be handled manually. */
+  title?: React.ReactNode;
   /** Text description of current progress value to display instead of percentage. */
   label?: React.ReactNode;
   /** Actual value of progress. */
@@ -34,7 +34,7 @@ export interface ProgressProps extends Omit<React.HTMLProps<HTMLDivElement>, 'si
   max?: number;
   /** Accessible text description of current progress value, for when value is not a percentage. Use with label. */
   valueText?: string;
-  /** Indicate whether to truncate the title */
+  /** Indicate whether to truncate the string title */
   isTitleTruncated?: boolean;
   /** Position of the tooltip which is displayed if title is truncated */
   tooltipPosition?: 'auto' | 'top' | 'bottom' | 'left' | 'right';