+++ packages/react-core/src/components/Select/Select.tsx
@@ -60,6 +60,10 @@ export interface SelectProps
   isDisabled?: boolean;
   /** Flag to indicate if the typeahead select allows new items */
   isCreatable?: boolean;
+  /** Flag indicating if placeholder styles should be applied */
+  hasPlaceholderStyle?: boolean;
+  /** @beta Flag indicating if the creatable option should set its value as a SelectOptionObject */
+  isCreateSelectOptionObject?: boolean;
   /** Value to indicate if the select is modified to show that validation state.
    * If set to success, select will be modified to indicate valid state.
    * If set to error, select will be modified to indicate error state.
@@ -140,7 +144,7 @@ export interface SelectProps
   chipGroupProps?: Omit<ChipGroupProps, 'children' | 'ref'>;
   /** Optional props to render custom chip group in the typeaheadmulti variant */
   chipGroupComponent?: React.ReactNode;
-  /** @beta Flag for retaining keyboard-entered value in typeahead text field when focus leaves input away */
+  /** Flag for retaining keyboard-entered value in typeahead text field when focus leaves input away */
   isInputValuePersisted?: boolean;
   /** Content rendered in the footer of the select menu */
   footer?: React.ReactNode;
@@ -180,6 +184,7 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
     isGrouped: false,
     isPlain: false,
     isDisabled: false,
+    hasPlaceholderStyle: false,
     isCreatable: false,
     validated: 'default',
     'aria-label': '',
@@ -211,7 +216,8 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
     favoritesLabel: 'Favorites',
     ouiaSafe: true,
     chipGroupComponent: null,
-    isInputValuePersisted: false
+    isInputValuePersisted: false,
+    isCreateSelectOptionObject: false
   };
 
   state: SelectState = {
@@ -343,7 +349,16 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
   updateTypeAheadFilteredChildren = (typeaheadInputValue: string, e: React.ChangeEvent<HTMLInputElement> | null) => {
     let typeaheadFilteredChildren: any;
 
-    const { onFilter, isCreatable, onCreateOption, createText, noResultsFoundText, children, isGrouped } = this.props;
+    const {
+      onFilter,
+      isCreatable,
+      onCreateOption,
+      createText,
+      noResultsFoundText,
+      children,
+      isGrouped,
+      isCreateSelectOptionObject
+    } = this.props;
 
     if (onFilter) {
       /* The updateTypeAheadFilteredChildren callback is not only called on input changes but also when the children change.
@@ -418,13 +433,24 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
       const newValue = typeaheadInputValue;
       if (
         !typeaheadFilteredChildren.find(
-          (i: React.ReactElement) => i.props.value.toLowerCase() === newValue.toLowerCase()
+          (i: React.ReactElement) =>
+            i.props.value && i.props.value.toString().toLowerCase() === newValue.toString().toLowerCase()
         )
       ) {
+        const newOptionValue = isCreateSelectOptionObject
+          ? ({
+              toString: () => newValue,
+              compareTo: value =>
+                this.toString()
+                  .toLowerCase()
+                  .includes(value.toString().toLowerCase())
+            } as SelectOptionObject)
+          : newValue;
+
         typeaheadFilteredChildren.push(
           <SelectOption
             key={`create ${newValue}`}
-            value={newValue}
+            value={newOptionValue}
             onClick={() => onCreateOption && onCreateOption(newValue)}
           >
             {createText} "{newValue}"
@@ -719,6 +745,7 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
       isGrouped,
       isPlain,
       isDisabled,
+      hasPlaceholderStyle,
       validated,
       selections: selectionsProp,
       typeAheadAriaLabel,
@@ -759,6 +786,7 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
       favoritesLabel,
       footer,
       loadingVariant,
+      isCreateSelectOptionObject,
       ...props
     } = this.props;
     const {
@@ -770,6 +798,11 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
     } = this.state;
     const selectToggleId = toggleId || `pf-select-toggle-id-${currentId++}`;
     const selections = Array.isArray(selectionsProp) ? selectionsProp : [selectionsProp];
+    // Find out if the selected option is a placeholder
+    const selectedOption = React.Children.toArray(children).find(
+      (option: any) => option.props.value === selections[0]
+    ) as any;
+    const isSelectedPlaceholder = selectedOption && selectedOption.props.isPlaceholder;
     const hasAnySelections = Boolean(selections[0] && selections[0] !== '');
     const typeaheadActiveChild = this.getTypeaheadActiveChild(typeaheadCurrIndex);
     let childPlaceholderText = null as string;
@@ -1029,6 +1062,9 @@ export class Select extends React.Component<SelectProps & OUIAProps, SelectState
           {...(footer && { footerRef: this.footerRef })}
           isOpen={isOpen}
           isPlain={isPlain}
+          hasPlaceholderStyle={
+            hasPlaceholderStyle && (!selections.length || selections[0] === null || isSelectedPlaceholder)
+          }
           onToggle={this.onToggle}
           onEnter={this.onEnter}
           onClose={this.onClose}